requirements.txt
streamlit
langchain
openai
chromadb
python-dotenv